Rework saving of private storage
This should make it more resilient as it does not
depend on an invisible box to hold the player's
storage items while the player is in game.

Instead the player's storage items are held in a variable
assocaited with the player, and de-serialized and serialzied when the
player accesses, and closes the storage respectively.