Fix #52
Save storage magazines in two separate arrays, i.e. PartialMagazines, and
FullMagazines.

The old "Magazines" property is still read for legacy reasons, but not
used anymore for saving. Eventually, we'll want to remove the reading part
as well.

Also, did a bit of refactoring to the order in which libraries are loaded.